Simplify L1 Index: key → file_id list (no versions)
Change L1 from key → [(version, file_id), ...] to key → [file_id, ...]
where file_ids are reverse-sorted by version (latest first). Version
resolution is now delegated to L2 indexes.

- put(key, file_id) prepends if not already at front
- getLatest(key) returns file_ids[0], O(1)
- getFileIds(key) returns full list for version lookups via L2
- removeFile(key, file_id) replaces removeOldVersions
- KeyRecord shrinks: vector<uint32_t> instead of vector<IndexEntry>
- DB::getByVersion now iterates L2 indexes for version resolution
- Snapshot format bumped to v2 (file_ids only, no versions)

+++ src/internal/log_entry.h
@@ -54,12 +54,11 @@ struct LogEntry {
     static constexpr size_t kChecksumSize = 4;
 };
 
-// Index entry stored in L1 and L2 indices.
-// L1: location = file_id (which log file)
-// L2: location = offset (byte offset within log file)
+// Index entry stored in the L1 index.
+// location = file_id (which log file), 32 bits is sufficient.
 struct IndexEntry {
     uint64_t version = 0;
-    uint64_t location = 0;
+    uint32_t location = 0;
 
     bool operator<(const IndexEntry& other) const {
         return version < other.version;
